% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra.R
\name{gra}
\alias{gra}
\title{Wykonanie raz gry w SuperFarmera}
\usage{
gra(strategia)
}
\arguments{
\item{strategia}{Funckja implementujaca strategie, ktora przyjmuje gracz podczas gry.}
}
\value{
liczba rzutow kostkami wykonanych do wygranej
}
\description{
Funkcja gra przeprowadza gre w SuperFarmera jeden raz przy zadanej
jako parametr strategii.  Zlicza ilosc rzutow kostkami, potrzebnych
aby wygrac. Jest jeden gracz.
}

