% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra_nk.R
\name{gra_nk}
\alias{gra_nk}
\title{Wykonanie raz gry w SuperFarmera}
\usage{
gra_nk(strategia)
}
\arguments{
\item{strategia}{Funckja implementujaca strategie, ktora przyjmuje gracz podczas gry.}
}
\value{
Liczba rzutow kostkami wykonanych do wygranej.
}
\description{
Funkcja gra przeprowadza gre w Super Farmera jeden raz przy zadanej
jako parametr strategii.  Zlicza ilosc rzutow kostkami potrzebnych,
aby wygrac. Jest jeden gracz.
}

