% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kup_zwierze_yolo.R
\name{kup_zwierze_yolo}
\alias{kup_zwierze_yolo}
\title{Funkcja pomocnicza strategii YOLO}
\usage{
kup_zwierze_yolo(jakie_zwierze, wektor_gracza)
}
\arguments{
\item{jakie_zwierze}{Liczba od 1 do 7, okreœlaj¹ca, które zwierze chcemy kupiæ.}

\item{wektor_gracza}{Wektor reprezentuj¹cy stan stada gracza}
}
\description{
Funkcja kup_zwierze_yolo dla danego jako parametr zwierzêcia kupuje je i zwraca wektor gracza
przy za³o¿eniu, ¿e zakup jest mo¿liwy.
}

