% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruch_gracz_anty_yolo.R
\name{ruch_gracz_anty_yolo}
\alias{ruch_gracz_anty_yolo}
\title{Ruch strategi¹ ANTY-YOLO}
\usage{
ruch_gracz_anty_yolo(stan_gracza)
}
\arguments{
\item{stan_gracza}{Wektor bêd¹cy stanem stada gracza}
}
\description{
Funkcja ruch_gracza_anty_yolo dla zadanego wektora gracza zwraca nowy wektor gracza po wykonaniu 
ruchu strategi¹ ANTY-YOLO, która skupia siê w pierwszej kolejnoœci na zakupie obu psów obronnych.
Jeœli ju¿ to osi¹gnie, grana jest strategia YOLO.
}

