% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruch_gracz_anty_yolo_zakup_duzego_psa.R
\name{ruch_gracz_anty_yolo_zakup_duzego_psa}
\alias{ruch_gracz_anty_yolo_zakup_duzego_psa}
\title{Zakup du¿ego psa}
\usage{
ruch_gracz_anty_yolo_zakup_duzego_psa(stan_gracza)
}
\arguments{
\item{stan_gracza}{Wektor bêd¹cy stanem stada gracza}
}
\description{
Funkcja pomocnicza strategii ANTY-YOLO, kupuj¹ca du¿ego psa obronnego, jeœli nas staæ
}

