% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruch_gracz_anty_yolo_zakup_malego_psa.R
\name{ruch_gracz_anty_yolo_zakup_malego_psa}
\alias{ruch_gracz_anty_yolo_zakup_malego_psa}
\title{Zakup ma³ego psa}
\usage{
ruch_gracz_anty_yolo_zakup_malego_psa(stan_gracza)
}
\arguments{
\item{stan_gracza}{Wektor bêd¹cy stanem stada gracza}
}
\description{
Funkcja pomocnicza strategii ANTY-YOLO, kupuj¹ca ma³ego psa obronnego, jeœli nas staæ
}

