% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruch_gracz_anty_yolo_zakup_psow.R
\name{ruch_gracz_anty_yolo_zakup_psow}
\alias{ruch_gracz_anty_yolo_zakup_psow}
\title{Zakup brakuj¹cych psów obronnych}
\usage{
ruch_gracz_anty_yolo_zakup_psow(stan_gracza)
}
\arguments{
\item{stan_gracza}{Wektor bêd¹cy stanem stada gracza}
}
\description{
Funkcja pomocnicza strategii ANTY-YOLO, kupuj¹ca brakuj¹cego obronnego, jeœli nas staæ
}

