% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzut_jedna_kostka.R
\name{rzut_jedna_kostka}
\alias{rzut_jedna_kostka}
\title{Rzut jedn¹ kostk¹}
\usage{
rzut_jedna_kostka(wektor_wartosci_kostki)
}
\arguments{
\item{wektor_wartosci_kostki}{}
}
\description{
Funkcja losuj¹ca z rozk³adu jednostajnego element wektora zadanego jako parametr
Tym parametrem jest wektor_czerwonej_kostki lub wektor_zielonej_kostki zdefiniowany w zbiorze danych kostki
}

