% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprawdz_czy_koniec.R
\name{sprawdz_czy_koniec}
\alias{sprawdz_czy_koniec}
\title{Funkcja sprawdz_czy_koniec}
\usage{
sprawdz_czy_koniec(x)
}
\arguments{
\item{x}{wektor bêd¹cy stanem stada gracza}
}
\description{
Funkcja sprawdzaj¹ca czy stan stada gracza spe³nia warunki zakoñczenia gry
}

