% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_nk.R
\name{strategia_nk}
\alias{strategia_nk}
\title{Strategia gry w Super Farmera}
\usage{
strategia_nk(herd)
}
\arguments{
\item{herd}{Obecny stan stada gracza.}
}
\value{
Stan stada po wymianach.
}
\description{
Funkcja strategia dokonuje wymian zwierzat na podstawie obecnego stanu stada.
}

