% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_yolo.R
\name{strategia_yolo}
\alias{strategia_yolo}
\title{Strategia YOLO}
\usage{
strategia_yolo(wektor_gracza)
}
\arguments{
\item{wektor_gracza}{Wektor bêd¹cy stanem stada gracza}
}
\value{
Wektor bêd¹cy stanem stada gracza po wykonaniu ruchu
}
\description{
Funkcja wykonuj¹ca ruch strategi¹ YOLO, która w ka¿dym ruchu stara siê kupiæ najdro¿sze brakuj¹ce w stadzie zwierze.
Zakup psów jest ignorowany
}

