% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zakup_zwierzecia.R
\name{zakup_zwierzecia}
\alias{zakup_zwierzecia}
\title{Zakup zwierzêcia}
\usage{
zakup_zwierzecia(stan_gracza, zwierzak)
}
\arguments{
\item{stan_gracza}{Wektor bêd¹cy stanem stada gracza}
}
\value{
Wektor gracza po udanej lub nie próbie zakupu
}
\description{
Funkcja pomocnicza strategii ANTY-YOLO, kupuj¹ca zadane w parametrze zwierze, jeêli nas staæ
}

