% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra_nk.R
\name{gra_nk}
\alias{gra_nk}
\title{Wykonanie raz gry w SuperFarmera}
\usage{
gra_nk(strategia)
}
\arguments{
\item{strategia}{Funckja implementująca strategię, którą przyjmuje gracz podczas gry.}
}
\value{
liczba rzutów kostkami wykonanych do wygranej
}
\description{
Funkcja gra przeprowadza grę w SuperFarmera jedne raz przy zadanej
jako parametr strategii.  Zlicza ilość rzutów kostkalimi, potrzebnych
aby wygrać. Jest jeden gracz.
}

