% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_nk.R
\name{strategia_nk}
\alias{strategia_nk}
\title{Strategia gry w SuperFarmera}
\usage{
strategia_nk(herd)
}
\arguments{
\item{herd}{Obecny stan stada gracza.}
}
\value{
stan stada po wymianach
}
\description{
Funkcja strategia dokonuje wymian zwierząt na podstawie obecnego stanu stada.
}

